+++ R/ms_drive_item.R
@@ -26,7 +26,7 @@
 #' - `get_parent_folder()`: Get the parent folder for this item, as a drive item object. Returns the root folder for the root. Not supported for remote items.
 #' - `get_path()`: Get the absolute path for this item, as a character string. Not supported for remote items.
 #' - `is_folder()`: Information function, returns TRUE if this item is a folder.
-#' - `load_dataframe(delim=NULL, ...)`: Download a delimited file and return its contents as a data frame. See 'Saving and loading data' below.
+#' - `load_dataframe(delim=NULL, ...)`: Download an Excel or text file and return its contents as a data frame. See 'Saving and loading data' below.
 #' - `load_rds()`: Download a .rds file and return the saved object.
 #' - `load_rdata(envir)`: Load a .RData or .Rda file into the specified environment.
 #' - `save_dataframe(df, file, delim=",", ...)` Save a dataframe to a delimited file.
@@ -81,8 +81,8 @@
 #'
 #' @section Saving and loading data:
 #' The following methods are provided to simplify the task of loading and saving datasets and R objects.
-#' - `load_dataframe` downloads a delimited file and returns its contents as a data frame. The delimiter can be specified with the `delim` argument; if omitted, this is "," if the file extension is .csv, ";" if the file extension is .csv2, and a tab otherwise. If the readr package is installed, the `readr::read_delim` function is used to parse the file, otherwise `utils::read.delim` is used. You can supply other arguments to the parsing function via the `...` argument.
-#' - `save_dataframe` is the inverse of `load_dataframe`: it uploads the given data frame to a folder item. Specify the delimiter with the `delim` argument. The `readr::write_delim` function is used to serialise the data if that package is installed, and `utils::write.table` otherwise.
+#' - `load_dataframe` downloads an Excel (.xlsx or .xls extension) or text file and returns its contents as a data frame. Loading Excel files uses the `readxl::read_excel` function, and so requires that the readxl package is installed. For a text file, you can specify the delimiter with the `delim` argument; if omitted, this is "," if the file extension is .csv, ";" if the file extension is .csv2, and a tab otherwise. If the readr package is installed, the `readr::read_delim` function is used to parse text files, otherwise `utils::read.delim` is used. You can supply other arguments to these functions via the `...` argument.
+#' - `save_dataframe` is the inverse of `load_dataframe`: it uploads the given data frame to a folder item. Specify the delimiter with the `delim` argument. The `readr::write_delim` function is used to serialise the data if that package is installed, and `utils::write.table` otherwise. Note that unlike loading, saving to an Excel file is not supported.
 #' - `load_rds` downloads a .rds file and returns its contents as an R object. It is analogous to the base `readRDS` function but for OneDrive/SharePoint drive items.
 #' - `save_rds` uploads a given R object as a .rds file, analogously to `saveRDS`.
 #' - `load_rdata` downloads a .RData or .Rda file and loads its contents into the given environment. It is analogous to the base `load` function but for OneDrive/SharePoint drive items.
@@ -93,7 +93,7 @@
 #'
 #' Support in the underlying Graph API for OData queries is patchy. Not all endpoints that return lists of objects support filtering, and if they do, they may not allow all of the defined operators. If your filtering expression results in an error, you can carry out the operation without filtering and then filter the results on the client side.
 #' @seealso
-#' [`ms_graph`], [`ms_site`], [`ms_drive`]
+#' [`AzureGraph::ms_graph`], [`ms_site`], [`ms_drive`]
 #'
 #' [Microsoft Graph overview](https://learn.microsoft.com/en-us/graph/overview),
 #' [OneDrive API reference](https://learn.microsoft.com/en-us/graph/api/resources/onedrive?view=graph-rest-1.0)
@@ -421,6 +421,17 @@ public=list(
     {
         private$assert_is_file()
         ext <- tolower(tools::file_ext(self$properties$name))
+
+        if(ext %in% c("xls", "xlsx"))
+        {
+            if(!requireNamespace("readxl"))
+                stop("The readxl package must be installed to load an Excel spreadsheet")
+            infile <- paste0(tempfile(), ".", ext)
+            on.exit(unlink(infile))
+            self$download(dest=infile)
+            return(readxl::read_excel(infile, ...))
+        }
+
         if(is.null(delim))
         {
             delim <- if(ext == "csv") "," else if(ext == "csv2") ";" else "\t"
@@ -430,7 +441,7 @@ public=list(
         {
             con <- rawConnection(dat, "r")
             on.exit(try(close(con), silent=TRUE))
-            readr::read_delim(con, delim=delim)
+            readr::read_delim(con, delim=delim, ...)
         }
         else utils::read.delim(text=rawToChar(dat), sep=delim, ...)
     },
@@ -566,6 +577,23 @@ private=list(
         on.exit(close(src$con))
 
         fullpath <- private$make_absolute_path(dest)
+
+        # handle zero-length files correctly: cannot use resumable upload
+        if(src$size == 0)
+        {
+            # possible fullpath formats -> string to append:
+            # drives/xxx/root -> /content
+            # drives/xxx/root:/foo/bar -> :/content
+            # drives/xxx/items/yyy -> /content
+            # drives/xxx/items/yyy:/foo/bar -> :/content
+            op <- if(grepl(":/", fullpath))
+                paste0(fullpath, ":/content")
+            else paste0(fullpath, "/content")
+
+            res <- call_graph_endpoint(self$token, op, http_verb="PUT", body=raw(0))
+            return(invisible(ms_drive_item$new(self$token, self$tenant, httr::content(res))))
+        }
+
         # possible fullpath formats -> string to append:
         # drives/xxx/root -> /createUploadSession
         # drives/xxx/root:/foo/bar -> :/createUploadSession